Move converters.js to ~/lib/utils/
